+++ .env
@@ -1,0 +1,1 @@
+VARIABLE_NAME=variable_value
